Move `no_packet_loops` test to `tests/` directory
