+++ rust/headless-client/Cargo.toml
@@ -35,13 +35,11 @@ url = { version = "2.5.2", default-features = false }
 uuid = { version = "1.10", default-features = false, features = ["std", "v4", "serde"] }
 
 [dev-dependencies]
+hex-literal = "0.4.1"
 tempfile = "3.10.1"
 
 [target.'cfg(target_os = "linux")'.dev-dependencies]
 mutants = "0.0.3" # Needed to mark functions as exempt from `cargo-mutants` testing
-
-[target.'cfg(target_os = "windows")'.dev-dependencies]
-hex-literal = "0.4.1"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 dirs = "5.0.1"
